+++ src/app/component/visit-summary/advice/advice.component.ts
@@ -64,7 +64,7 @@ export class AdviceComponent implements OnInit {
     private diagnosisService: DiagnosisService,
     private snackbar: MatSnackBar,
     private route: ActivatedRoute
-  ) {}
+  ) { }
 
   search = (text$: Observable<string>) =>
     text$.pipe(
@@ -74,8 +74,8 @@ export class AdviceComponent implements OnInit {
         term.length < 1
           ? []
           : this.advices
-              .filter((v) => v.toLowerCase().indexOf(term.toLowerCase()) > -1)
-              .slice(0, 10)
+            .filter((v) => v.toLowerCase().indexOf(term.toLowerCase()) > -1)
+            .slice(0, 10)
       )
     );
 
@@ -108,7 +108,7 @@ export class AdviceComponent implements OnInit {
     const value = form.advice;
     const providerDetails = getFromStorage("provider");
     const providerUuid = providerDetails.uuid;
-    if (providerDetails && providerUuid === getEncounterProviderUUID()) {
+    if (this.diagnosisService.isSameDoctor()) {
       this.encounterUuid = getEncounterUUID();
       const json = {
         concept: this.conceptAdvice,
@@ -121,17 +121,15 @@ export class AdviceComponent implements OnInit {
         this.diagnosisService.isVisitSummaryChanged = true;
         this.advice.push({ uuid: response.uuid, value: value });
       });
-    } else {
-      this.snackbar.open("Another doctor is viewing this case", null, {
-        duration: 4000,
-      });
     }
   }
 
   delete(i) {
-    const uuid = this.advice[i].uuid;
-    this.diagnosisService.deleteObs(uuid).subscribe((res) => {
-      this.advice.splice(i, 1);
-    });
+    if (this.diagnosisService.isSameDoctor()) {
+      const uuid = this.advice[i].uuid;
+      this.diagnosisService.deleteObs(uuid).subscribe((res) => {
+        this.advice.splice(i, 1);
+      });
+    }
   }
 }
