+++ src/app/component/visit-summary/follow-up/follow-up.component.ts
@@ -65,7 +65,7 @@ export class FollowUpComponent implements OnInit {
     private snackbar: MatSnackBar,
     private route: ActivatedRoute,
     private datepipe: DatePipe
-  ) {}
+  ) { }
 
   ngOnInit() {
     this.visitUuid = this.route.snapshot.paramMap.get("visit_id");
@@ -88,7 +88,7 @@ export class FollowUpComponent implements OnInit {
     const advice = form.advice;
     const providerDetails = getFromStorage("provider");
     const providerUuid = providerDetails.uuid;
-    if (providerDetails && providerUuid === getEncounterProviderUUID()) {
+    if (this.diagnosisService.isSameDoctor()) {
       this.encounterUuid = getEncounterUUID();
       const json = {
         concept: this.conceptFollow,
@@ -101,17 +101,15 @@ export class FollowUpComponent implements OnInit {
         this.diagnosisService.isVisitSummaryChanged = true;
         this.followUp.push({ uuid: resp.uuid, value: json.value });
       });
-    } else {
-      this.snackbar.open("Another doctor is viewing this case", null, {
-        duration: 4000,
-      });
     }
   }
 
   delete(i) {
-    const uuid = this.followUp[i].uuid;
-    this.diagnosisService.deleteObs(uuid).subscribe((res) => {
-      this.followUp.splice(i, 1);
-    });
+    if (this.diagnosisService.isSameDoctor()) {
+      const uuid = this.followUp[i].uuid;
+      this.diagnosisService.deleteObs(uuid).subscribe((res) => {
+        this.followUp.splice(i, 1);
+      });
+    }
   }
 }
