+++ src/app/component/visit-summary/prescribed-medication/prescribed-medication.component.ts
@@ -79,7 +79,7 @@ export class PrescribedMedicationComponent implements OnInit {
     private diagnosisService: DiagnosisService,
     private snackbar: MatSnackBar,
     private route: ActivatedRoute
-  ) {}
+  ) { }
 
   searchPrescription = (text$: Observable<string>) =>
     text$.pipe(
@@ -89,8 +89,8 @@ export class PrescribedMedicationComponent implements OnInit {
         term.length < 1
           ? []
           : this.conceptPrescription
-              .filter((v) => v.toLowerCase().indexOf(term.toLowerCase()) > -1)
-              .slice(0, 10)
+            .filter((v) => v.toLowerCase().indexOf(term.toLowerCase()) > -1)
+            .slice(0, 10)
       )
     );
 
@@ -102,8 +102,8 @@ export class PrescribedMedicationComponent implements OnInit {
         term.length < 1
           ? []
           : this.conceptfrequency
-              .filter((v) => v.toLowerCase().indexOf(term.toLowerCase()) > -1)
-              .slice(0, 10)
+            .filter((v) => v.toLowerCase().indexOf(term.toLowerCase()) > -1)
+            .slice(0, 10)
       )
     );
 
@@ -115,8 +115,8 @@ export class PrescribedMedicationComponent implements OnInit {
         term.length < 1
           ? []
           : this.conceptAdministration
-              .filter((v) => v.toLowerCase().indexOf(term.toLowerCase()) > -1)
-              .slice(0, 10)
+            .filter((v) => v.toLowerCase().indexOf(term.toLowerCase()) > -1)
+            .slice(0, 10)
       )
     );
 
@@ -128,8 +128,8 @@ export class PrescribedMedicationComponent implements OnInit {
         term.length < 1
           ? []
           : this.conceptDose
-              .filter((v) => v.toLowerCase().indexOf(term.toLowerCase()) > -1)
-              .slice(0, 10)
+            .filter((v) => v.toLowerCase().indexOf(term.toLowerCase()) > -1)
+            .slice(0, 10)
       )
     );
 
@@ -141,8 +141,8 @@ export class PrescribedMedicationComponent implements OnInit {
         term.length < 1
           ? []
           : this.conceptDurationUnit
-              .filter((v) => v.toLowerCase().indexOf(term.toLowerCase()) > -1)
-              .slice(0, 10)
+            .filter((v) => v.toLowerCase().indexOf(term.toLowerCase()) > -1)
+            .slice(0, 10)
       )
     );
 
@@ -209,7 +209,7 @@ export class PrescribedMedicationComponent implements OnInit {
     }
     const providerDetails = getFromStorage("provider");
     const providerUuid = providerDetails.uuid;
-    if (providerDetails && providerUuid === getEncounterProviderUUID()) {
+    if (this.diagnosisService.isSameDoctor()) {
       this.encounterUuid = getEncounterUUID();
       const json = {
         concept: this.conceptMed,
@@ -223,17 +223,15 @@ export class PrescribedMedicationComponent implements OnInit {
         this.meds.push({ uuid: response.uuid, value: insertValue });
         this.add = false;
       });
-    } else {
-      this.snackbar.open("Another doctor is viewing this case", null, {
-        duration: 4000,
-      });
     }
   }
 
   delete(i) {
-    const uuid = this.meds[i].uuid;
-    this.diagnosisService.deleteObs(uuid).subscribe((res) => {
-      this.meds.splice(i, 1);
-    });
+    if (this.diagnosisService.isSameDoctor()) {
+      const uuid = this.meds[i].uuid;
+      this.diagnosisService.deleteObs(uuid).subscribe((res) => {
+        this.meds.splice(i, 1);
+      });
+    }
   }
 }
